Add component for rule creation and dropdown
